consumers/reference-solution-consumer/pyproject.toml
[project]
name = "reference-solution-consumer"
version = "0.1.0"
description = ""
authors = [
    {name = "parthokr",email = "partho.kr@proton.me"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "base-consumer (>=0.1.1,<0.2.0)",
    "requests (>=2.32.3,<3.0.0)",
    "grpcio (>=1.70.0,<2.0.0)",
    "grpcio-tools (>=1.70.0,<2.0.0)",
    "boto3 (>=1.37.9,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
]

[tool.poetry]
packages = [{include = "reference_solution_consumer", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
